refactor: move lark_app to notifiers directory and update imports
